hid parts of backstage
much of the backstage ui would be confusing for ilga.org users.
the confusing bits have been hidden for now.
This commit updates the login to point to the ilga.org challenger (this
should change to something more polished later).
The frontpage makes use of this rather than introducing another variable to worry
about hence the delete.