mangled link creation
In ILGA, it is important that we can link up articles of the same
name for translation purposes.

A French tiddler will have a different article title than an English
tiddler. Thus guids are used to link the two.

A heading field is used for the readable title. This code looks for a
heading field and uses it in place of the title / server.title field
ensuring that search results and any references to tiddlers display
human readable links.