modelling tiddlyspace concepts as tiddlers
this turns concepts such as inclusions, members, spaces belong to a
user and identities into tiddlers. These are tagged system-inclusions,
system-members, system-space, system-inclusion respectively (system followed
by the concept name).

the main method mapConcept takes a name, a concept (for generating the tag)
and an optional info string parameter that can be used to record data
associated with a concept. It is used at the moment just by the inclusion
mechanism which needs to record the order of included spaces.