re-implemented TiddlySpaceIdentities list
now when listing identities a simple invocation of the list macro
is made providing the filter [tag[system-identity]] and the template
with name TemplateTiddlySpaceAdmin (added as shadow tiddler)

the refresh function is made redundant.

A generic listConcept function is created with view of extending this
for other concepts other than identities.