implemented deleteConcept view type
This takes a given value on a tiddler and treats it as a concept
e.g. name of space / name of identity
The deleteConcept view type creates a delete button that when clicked
fires of the delete handler associated with that concept.

Note changes to TiddlySpaceMembers required for this to work.